backend/requirements.txt
uvicorn==0.30.6
fastapi==0.114.0
pyjwt==2.9.0
requests==2.32.3
cryptography==43.0.1
langchain==0.2.16
langchain-ollama==0.1.3
langchain-core==0.2.39
langsmith==0.1.118
sqlalchemy==2.0.34
passlib==1.7.4
python-multipart==0.0.10